added new project helper function
